+++ src/users/application/services/users.service.spec.ts
@@ -1,11 +1,10 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { UsersServiceImpl, UsersServiceSymbol } from './users.service.impl';
-import { UsersService } from './users.service';
-import { UserOutputDto } from '../dto/create-user.dto';
-import { UsersRepository } from '../repositories/users.repository';
-import { UsersRepositorySymbol } from '../repositories/users.repository.impl';
-import { User } from '../entities/user.entity';
-
+import { User } from 'src/users/domain/entities/user.entity';
+import { UsersService } from './user.service';
+import { UsersRepository } from 'src/users/domain/repositories/users.repository';
+import { UsersRepositorySymbol } from 'src/users/infrastructure/persistence/repositories/users.repository.impl';
+import { UserResponseDto } from 'src/users/presentation/dto/response/user.response.dto';
 const mockUsersRepository = {
   createUser: jest.fn(),
   getUser: jest.fn(),
@@ -48,11 +47,9 @@ describe('UsersService', () => {
 
   describe('유저 생성', () => {
     it('유저를 생성하면 해당 유저를 반환한다', async () => {
-   
-
       usersRepository.createUser.mockResolvedValue(user);
 
-      const res: UserOutputDto = await service.createUser({
+      const res: UserResponseDto = await service.createUser({
         name: user.name,
         email: user.email,
       });
