+++ src/users/domain/entities/user.entity.ts
