+++ src/users/presentation/dto/request/create-user.request.dto.ts
@@ -1,0 +1,9 @@
+import { IsDate, IsEmail, IsString, IsUUID } from 'class-validator';
+
+export class CreateUserDto {
+  @IsString()
+  name: string;
+
+  @IsEmail()
+  email: string;
+}
